Adding interruptions for call instrucctions

+++ SPARC/UART_Communication/Coordenadas/Comandos.X/cases.h
@@ -5,8 +5,8 @@ int coord(char* P1, char* L, unsigned short* x, unsigned short* y, char* P2);
 
 char Par1;
 char letter;
-unsigned short cord_x;
-unsigned short cord_y;
+unsigned int cord_x;
+unsigned int cord_y;
 char Par2;
 
 int x;
